backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
apify-client
selenium
webdriver-manager
beautifulsoup4
undetected-chromedriver
